update to Navigation panel
Comments and new functionalities have been added. The sub-buttons are
now collapsible. clicking on parent buttons will only show or hide
sub-buttons and won't redirect to a screen.
Also, a null can now be passed as a screen and will result in the button
not having an actionListnener.
Now, sub-button can be removed.
targeted screen can now be changed.
setPosition was overriden so that the panel will extend only downward.