Rename Company Review
Went over the Rename Company code and fixed some bugs/errors. Instead
of the user being able to click anywhere to bring up the text box to
enter the company name, he/she has to click on the button for it. Once
the button is clicked, the user can enter in the company name which is
then stored in the program. Also cleaned up the code so it can be
easier for the other people in my group to use it.